+++ guildwars2api/client.py
@@ -1,9 +1,7 @@
 import requests
 
-from guildwars2api.resources import (GuildDetails, EventNames, MapNames, WorldNames, Matches, MatchDetails, 
-                                     ObjectiveNames, Items, ItemDetails, Recipes, RecipeDetails, ItemsV2, 
-                                     RecipesV2, Skins, Skin_Details, SkinsV2)
-
+import guildwars2api.api_2 as v2
+import guildwars2api.api_1 as v1
 
 class GuildWars2API(object):
     """
@@ -27,39 +25,25 @@ class GuildWars2API(object):
         }
 
         if (api_version == 'v2'):
-            self.items = self._prepare(ItemsV2)
-            self.recipes = self._prepare(RecipesV2)
-            self.skins = self._prepare(SkinsV2)
+            self.items = self._prepare(v2.Items)
+            self.recipes = self._prepare(v2.Recipes)
+            self.skins = self._prepare(v2.Skins)
         elif (api_version =='v1'):
-            self.items = self._prepare(Items)
-            self.recipes = self._prepare(Recipes)
-            self.event_names = self._prepare(EventNames)
-            self.map_names = self._prepare(MapNames)
-            self.world_names = self._prepare(WorldNames)
-            self.matches = self._prepare(Matches)
-            self.match_details = self._prepare(MatchDetails)
-            self.objective_names = self._prepare(ObjectiveNames)
-            self.item_details = self._prepare(ItemDetails)
-            self.recipe_details = self._prepare(RecipeDetails)
-            self.guild_details = self._prepare(GuildDetails)
-            self.skins = self._prepare(Skins)
-            self.skin_details = self._prepare(Skin_Details)
-            # self.events = self._prepare(Events)
-        else: # Default to v1
-            self.items = self._prepare(Items)
-            self.recipes = self._prepare(Recipes)
-            self.event_names = self._prepare(EventNames)
-            self.map_names = self._prepare(MapNames)
-            self.world_names = self._prepare(WorldNames)
-            self.matches = self._prepare(Matches)
-            self.match_details = self._prepare(MatchDetails)
-            self.objective_names = self._prepare(ObjectiveNames)
-            self.item_details = self._prepare(ItemDetails)
-            self.recipe_details = self._prepare(RecipeDetails)
-            self.guild_details = self._prepare(GuildDetails)
-            self.skins = self._prepare(Skins)
-            self.skin_details = self._prepare(Skin_Details)
-            # self.events = self._prepare(Events)
+            self.items = self._prepare(v1.Items)
+            self.recipes = self._prepare(v1.Recipes)
+            self.recipe_details = self._prepare(v1.RecipeDetails)
+            self.item_details = self._prepare(v1.ItemDetails)
+            self.skins = self._prepare(v1.Skins)
+            self.skin_details = self._prepare(v1.Skin_Details)
+            self.matches = self._prepare(v1.Matches)
+            self.match_details = self._prepare(v1.MatchDetails)
+            self.objective_names = self._prepare(v1.ObjectiveNames)
+            self.event_names = self._prepare(v1.EventNames)
+            self.map_names = self._prepare(v1.MapNames)
+            self.world_names = self._prepare(v1.WorldNames)
+            self.guild_details = self._prepare(v1.GuildDetails)
+        else:
+            raise ValueError("Only v1 or v2 accepted")
     
     def _prepare(self, resource):
         return resource(self.options, self.session)
